Add the public interface of the CardMatchingGame class
initWithCardCount:usingDeck: is our designated initializer for which we specify the number of cards we want to play with and the card deck to use.
flipCardAtIndex: flips a card at a specified index.
cardAtIndex: returns a card that can be used for displaying the UI for example.
score is a readonly propriety that internally keeps the score.