Implement flipCardAtIndex: method from the CardMatchingGame class
The whole logic of the game is in this method. When we flip a card up we check other cards that are up, and if they match we increment the score by a defined constant, otherwise we decrement the score. We also decrement the score by a constant amount every time we flip a card up.